+++ firebase.json
@@ -12,5 +12,28 @@
         "destination": "/index.html"
       }
     ]
+  },
+  "emulators": {
+    "auth": {
+      "port": 9099,
+      "data": "./.firebase/emulator-auth-data"
+    },
+    "firestore": {
+      "port": 8089,
+      "rules": "firestore.rules",
+      "data": "./.firebase/emulator-firestore-data"
+    },
+    "storage": {
+      "port": 9199,
+      "rules": "storage.rules",
+      "data": "./.firebase/emulator-storage-data"
+    },
+    "ui": {
+      "enabled": true
+    }
+  },
+  "singleProjectMode": true,
+  "storage": {
+    "rules": "storage.rules"
   }
 }
