+++ index.html
@@ -154,6 +154,6 @@
 <body class="text-slate-800 dark:text-slate-200 font-sans bg-light-bg dark:bg-slate-900">
   <div id="page-background"></div>
   <div id="root"></div>
-  <script type="module" src="/index.tsx"></script>
+  <script type="module" src="./src/index.tsx"></script>
 </body>
 </html>
