+++ src/components/common/FullscreenImageViewer.tsx
