+++ src/components/common/ImageBankSettingsModal.tsx
