+++ src/components/common/SphereDisplay.tsx
