+++ src/pages/ImageBankPage.tsx
@@ -9,7 +9,7 @@ import { LoadingSpinner } from '../components/common/LoadingSpinner';
 import { getUserById } from '../services/userService'; // Import for getting user details
 import ExifReader from 'exifreader';
 import { getDownloadURL, ref } from 'firebase/storage'; // Added
-import { storage } from '../firebase'; // Added
+import { storage } from '../../firebase'; // Added
 // Removed: import { generateImageBankExportPdf } from '../services/pdfService'; 
 
 interface ImageBankPageProps {
