+++ src/pages/auth/EmailConfirmationPage.tsx
