+++ src/services/geminiService.ts
@@ -1,7 +1,7 @@
 /// <reference path="../types.ts" />
 
 import { GoogleGenAI, GenerateContentResponse, Part } from "@google/genai";
-import { GEMINI_TEXT_MODEL } from '../constants'; // GEMINI_IMAGE_MODEL was same as text, using text directly
+import { GEMINI_TEXT_MODEL } from "../constants"; // GEMINI_IMAGE_MODEL was same as text, using text directly
 import { UserDescriptionEntry } from "../types";
 
 // Strict adherence to guideline: API key MUST be obtained from environment
