+++ src/services/pdfService.ts
