+++ tsconfig.json
@@ -23,8 +23,9 @@
     "noFallthroughCasesInSwitch": true,
     "noUncheckedSideEffectImports": true,
 
+    "baseUrl": "src",
     "paths": {
-      "@/*" :  ["./*"]
+      "@/*": ["./*"]
     }
   }
 }
